// src/api/generated/models/productMediaResponseDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { ProductMediaResponseDtoType } from './productMediaResponseDtoType';

export interface ProductMediaResponseDto {
  /**
     * @minLength 26
     * @maxLength 26
     * @pattern ^[0-7][0-9A-HJKMNP-TV-Z]{25}$
     */
  id: string;
  /**
     * @minLength 26
     * @maxLength 26
     * @pattern ^[0-7][0-9A-HJKMNP-TV-Z]{25}$
     */
  productId: string;
  /**
     * @minLength 26
     * @maxLength 26
     * @nullable
     * @pattern ^[0-7][0-9A-HJKMNP-TV-Z]{25}$
     */
  variantId: string | null;
  type: ProductMediaResponseDtoType;
  url: string;
  /** @nullable */
  altText: string | null;
  sortOrder: number;
  isPrimary: boolean;
  createdAt: string;
  updatedAt: string;
}